requirements.txt
matplotlib==3.5.0
pandas==1.3.4
Pillow==8.4.0
PyYAML==6.0
torch==1.10.0
torchvision==0.11.1
transformers==4.30.2
tqdm==4.66.1
